This is my first JavaScript snippet:
onload=init Makes that the function init() gets executed on page
loading.

document.getElementsByTagName('tag') This is used to get all the
elements declared with the same tag in the page that called the script.
It returns an array of objects that I stored in the variable h1tags.

onclick This runs what's after the = after that the user has clicked the
object.

Explaining the changeColor() function:
this is a reference to the object that called the function.

innerHTML is used to change the text of the object but I need to learn
more about it.

randomColor is a variable that stores a random hex color value generated
using some math functions and then wrapped into a string.

style.color is the color value of the object, that will change to
randomColor value.

Looks nice :p