Get a really basic CSS style:
color inside h1{} changes initial color of all the h1 tags in the
document that references this style.

a:$state is used to change the color of a link depending on his state.
the a:not(visited) is used to change the color of a link which is NOT
visited.

body{} is used to change body properties, in this case the background.

div#wrapper{} is used to change the properties of a div with the id
wrapper.